burrTxt2 option to select all problems
Adds "-B" option which will solve all problems in a given puzzle file,
not just the first (default) or a single one (specified by "-b").

This also clarifies a the variable name `lastProblem`. Previously the
name was misleading, since it was actually one more than the last
problem to be solved. Now it's actually the index of the last problem to
be solved.