Add compile-time switch to use recursive assembler
assembler_1 had an alternative commented out implementation that is
recursive instead of an iterative task stack. The intention that the
recursive implementation equivalent but easier to understand, and can be
used as a starting point in understanding the iterative version.

This commit adds an option in `assembler_1.h` to switch to the recursive
implementation at compile time. It works the same out of the box as far
as I can tell, it's just slower. I only had to fix a few minor compiler
warnings to get it to compile.