Limit solve drop/limit to be positive to fix crash
Manually inputting a zero or negative number in the solver's "Drop" or
"Limit" paramaters can result in a crash or infinite loop, so this
change limits their values to be strictly positive (at least 1). It
would be ideal to also make the UI limit the user input, not just on the
backend in solvethread, but I didn't find a simple way to do that.

The particular crashes this fixes are:

* Setting drop = 0 causes a divide by zero in lib/solvethread.cpp:155
* Setting drop = negative or limit = 0 causes an infinite loop in lib/solvethread.cpp:363