Prevent click event every movement in 3d view
This solves the issue of clicking (with ctrl, shift or alt) to add or
remove voxels, but accidently moving the mouse slightly, causing way
more voxels to be added or removed then intended. Every mouse move event
would trigger another click event, which quickly does something the user
likely didn't mean to.

This was originally intended behavior (see
8782ebd30cc99323257ff4fe5702b0da56d2986c) meant to make it "much more
easy to 'drill' holes". However, I think it causes a bigger problem than
it solves. The use case of rapidly adding or removing voxels isn't that
common and it's not a controlled way to do it anyways. Doing this on
accident is much to easy.