Disable -Woverloaded-virtual error
The problem is that `voxel_c::getMesh()` is a virtual method, but it has
a different signature in `voxel_2_c`. This isn't a problem in practice
because the the calling code will always call the appropriate method
depending on the subclass (`stlExporter_2_c` for example calls the
correct signature for `voxel_2_c`).

Actually getting rid of the compiler warning without disabling it will
require more effort due to how the code is currently organized.
`voxel_c` has a default implementation for `getMesh()` and
`getMeshInternal()`, so we can't simply take those out of the base
class. The stl exporters are't actually passed the subclass type, only
the base class, which would also have to be changed.

So that's why I opted to fix this by adding a compiler flag to ignore
"overloaded-virtual". Older compilers didn't have this problem because
this check wasn't turned on by default. At least on my local machine,
gcc 10.2.1 has the option even though it's not documented or turned on
by "-Wall -Wextra".