Fix dangling pointer error
Fixes error in more recent versions of GCC with -Wdangling-pointer.
Without this fix, this error occurs:

    lib/stl.cpp:148:65: error: dangling pointer to an unnamed temporary may be used [-Werror=dangling-pointer=]
    148 |         fprintf(f,"  facet normal %9.4e %9.4e %9.4e\n", normal[0], normal[1], normal[2]);
        |                                                         ~~~~~~~~^
    lib/stl.cpp:115:38: note: unnamed temporary defined here
    115 |     const float * normal = fc->normal().getData();
        |                            ~~~~~~~~~~^~

This happens because the scope of the value `fc->normal()` is only
within that one line. It's an "unnamed temporary". Using a pointer from
it later, i.e. `fc->normal().getData()` which is used later in
`normal[0]`, isn't valid.