Initial undo / redo functionality
There are still some rough edges, but the functionality is there. Undo
history is saved in XML format for now, since that was easiest, but we
may need to do something more intelligent in the future. Other than
that, most of the rough edges are with the user interface, such as
keeping the selected entity / problem while undoing, and the keyboard
shortcuts not working on the problems tab.