Disable undo/redo menu items when cannot undo/redo
Disables the menu item under the file menu for "Undo" and "Redo" when
there is nothing to undo or redo respectively. This also disables the
keyboard shortcuts, since the menu currently handles that.

The implementation is straightforward but there are some ordering issues
in which `updateInterface()` needs to be called after
`undoManager->recordState()`, now that the former depends on undo state
recorded by the latter. There were both places where `updateInterface()`
wasn't called, and where it was called in the wrong order.