Disable edits while solving to prevent segfaults
Most edits are already disabled during solving because editing shapes
the solver is using can cause malformed solutions or crashes in the
solver thread. There are a few places that were missed though, which
this change addresses.

The particular callbacks in `mainWindow_c` which we disable during
solving are:

* `cb_3dClick()`: Ctrl-click on the 3D view to add voxels. Malformed solutions or segfault occur if shapes the solver was using are edited.
* `cb_StatusWindow()`: The status menu item opens a dialog which can delete shapes. A segfault occurs if shapes the solver was using are deleted.
* `cb_Convert()`: Use "Convert" under the "File" menu to change grids. Segfault occurs.

To reproduce these issues, insert a `sleep(10)` somewhere in
`solveThread_c::run()`. Alternatively, test with a puzzle which takes
long enough to solve that you can edit the puzzle before it finishes.
Start the puzzle solving then edit the puzzle and the issue will occur
when the solving is finished.