Increase solve timer precision
This changes solvethread from using `time()`, with second precision, to
using `std::chrono::steady_clock`, which should be higher precision,
plus it's monotonic.

It's necessary to perform our own timing within solvethread for
benchmark measurements because burrTxt2 uses polling to see if
solvethread is finished, so timing burrTxt2 execution isn't a precise
way to measure solve time. This change will allow us to perform proper
benchmarks on optimizations, both current and future. In the future it
would also be useful to have timing information for each individual
step of the solve process.